feat : 차트category - color 연결
