+++ package.json
@@ -15,6 +15,7 @@
     "@mui/material": "^5.10.1",
     "@tailwindcss/forms": "^0.5.2",
     "apexcharts": "^3.35.5",
+    "axios": "^0.27.2",
     "moment": "^2.29.4",
     "next": "12.2.5",
     "next-auth": "^4.10.3",
