+++ src/components/charts/Charts.tsx
@@ -1,13 +1,16 @@
 import { donutOptions, barOptions } from '@application/chart/Options';
 import { seriesData } from '@application/chart/Series';
 import { MOCK_STATISTICS } from '@helper/mock';
-import { getStaticsDay } from '@infra/chart/chartApi';
+import { fetchStat1Day, getStaticsDay } from '@infra/chart/chartApi';
 import { Box } from '@mui/material';
+import axios from 'axios';
 import dynamic from 'next/dynamic';
+import useSWR from 'swr';
 const ApexChart = dynamic(() => import('react-apexcharts'), { ssr: false });
 
 export const Charts = () => {
-  const data = getStaticsDay({ mode: '1day', oneDay: '20220910' });
+  const fetcher = (url: string) => axios.get(url).then(res => res.data);
+  const { data, error } = useSWR('stat/allday', fetcher);
   console.log(data);
   return (
     <Box
