+++ src/components/diary/index.tsx
@@ -1,8 +1,8 @@
-import { Box, Typography, styled } from '@mui/material';
+import { DayEmotion, getDayEmotion } from '@application/diary/DayResults';
+import { Box, styled } from '@mui/material';
 import moment from 'moment';
 import dynamic from 'next/dynamic';
 import { useEffect, useState } from 'react';
-import { DayEmotion, getDayEmotion } from '../../application/diary/DayResults';
 import { Emotion } from './Emotion';
 const Calendar = dynamic(() => import('react-calendar'), { ssr: false });
 
@@ -11,16 +11,6 @@ const Calendar = dynamic(() => import('react-calendar'), { ssr: false });
  * 1. Calendar Css 수정 필요 (img 들어가게)
  * 2. 서버로 부터 해당 일에 대한 감정, 채팅 내역을 불러 오게
  */
-
-//
-const StyleCalendar = styled(Calendar)`
-  display: flex;
-  border-radius: 10px;
-  width: 480px;
-  height: 960px;
-  border: 1px solid black;
-`;
-
 const Diary = () => {
   const [value, setDate] = useState(new Date());
   const [emotions, setEmotions] = useState<DayEmotion[]>([]);
@@ -50,7 +40,6 @@ const Diary = () => {
           );
         }}
       />
-      <Typography color={'GrayText'}> {value.toDateString()}</Typography>
     </Box>
   );
 };
