+++ src/pages/calendar/index.tsx
@@ -1,6 +1,5 @@
 import type { NextPage } from 'next';
 import { Container } from '@mui/material';
-import CharacterContainer from '../../components/character';
 import Diary from '../../components/diary';
 import { Charts } from '../../components/charts/Charts';
 
@@ -11,11 +10,9 @@ const CalendarPage: NextPage = () => {
         display: 'flex',
         flexDirection: 'column',
         alignItems: 'center',
-        mt: 2,
         height: '100vh',
       }}
     >
-      <CharacterContainer />
       <Diary />
       <Charts />
     </Container>
